functional status as of 20.03.2021
this version will currently not change personal PW for testing reason.
Login will work through djangi admin DB but should change to osmapi DB
in one of the next steps.
It is possible to create new vessels with all kind of parameters, and
one can display own Track listing. Track upload is not part of this
Version.